chore: remove selected File view
